+++ tests/testthat/test-precipitation_upper.R
@@ -13,24 +13,24 @@ ofile      <- NA
 delta.file <- NA
 
 test_that("2030 decadal prediction", {
-  sc="GL"
+  scenario="GL"
 
   horizon = 2030
   tmp <- TransformPrecip(ifile=ifile,
                                           ofile=ofile,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon,
                                           subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14___2030_rr_upper.rds")
 })
 
 test_that("Scenario GL", {
-  sc="GL"
+  scenario="GL"
 
   horizon = 2050
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_GL_2050_rr_upper.rds")
@@ -38,7 +38,7 @@ test_that("Scenario GL", {
   horizon = 2085
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_GL_2085_rr_upper.rds")
@@ -46,12 +46,12 @@ test_that("Scenario GL", {
 
 
 test_that("Scenario GH", {
-  sc="GH"
+  scenario="GH"
 
   horizon = 2050
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_GH_2050_rr_upper.rds")
@@ -59,19 +59,19 @@ test_that("Scenario GH", {
   horizon = 2085
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_GH_2085_rr_upper.rds")
 })
 
 test_that("Scenario WH", {
-  sc="WH"
+  scenario="WH"
 
   horizon = 2050
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_WH_2050_rr_upper.rds")
@@ -79,19 +79,19 @@ test_that("Scenario WH", {
   horizon = 2085
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_WH_2085_rr_upper.rds")
 })
 
 test_that("Scenario WL", {
-  sc="WL"
+  scenario="WL"
 
   horizon = 2050
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_WL_2050_rr_upper.rds")
@@ -99,7 +99,7 @@ test_that("Scenario WL", {
   horizon = 2085
   tmp <- TransformPrecip(ifile=ifile,
                                        ofile=ofile,
-                                       sc=sc,
+                                       scenario=scenario,
                                        horizon = horizon,
                                        subscenario=subscenario)
   expect_equal_to_reference(tmp, "regressionOutput/precipitation/KNMI14_WL_2085_rr_upper.rds")