+++ tests/testthat/test-radiation.R
@@ -12,25 +12,25 @@ ofile      <- NA
 delta.file <- NA
 
 test_that("2030 decadal prediction", {
-  sc="GL"
+  scenario="GL"
 
   horizon = 2030
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile="tmp.txt",
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14___2030_rsds.rds")
 })
 
 test_that("Scenario GL", {
-  sc="GL"
+  scenario="GL"
 
   horizon = 2050
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_GL_2050_rsds.rds")
 
@@ -38,19 +38,19 @@ test_that("Scenario GL", {
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_GL_2085_rsds.rds")
 })
 
 test_that("Scenario GH", {
-  sc="GH"
+  scenario="GH"
 
   horizon = 2050
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_GH_2050_rsds.rds")
 
@@ -58,46 +58,46 @@ test_that("Scenario GH", {
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_GH_2085_rsds.rds")
 })
 
 test_that("Scenario WH 2050", {
-  sc="WH"
+  scenario="WH"
 
   horizon = 2050
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_WH_2050_rsds.rds")
 })
 
 test_that("Scenario WH 2085", {
   #skip_on_travis()
-  sc="WH"
+  scenario="WH"
   horizon = 2085 # This set is not produced by this version of the code - neither with
   # a factor of 0.7 or 0.75 with old delta file it works
   oldDeltaFile = "regressionInput/radiation/deltas-KNMI14__rsds_WH__2085_v201504.txt"
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=oldDeltaFile,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
 
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_WH_2085_rsds.rds")
 })
 
 test_that("Scenario WL", {
-  sc="WL"
+  scenario="WL"
 
   horizon = 2050
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_WL_2050_rsds.rds")
 
@@ -105,7 +105,7 @@ test_that("Scenario WL", {
   tmp <- TransformRadiation(ifile=ifile,
                                           ofile=ofile,
                                           delta.file=delta.file,
-                                          sc=sc,
+                                          scenario=scenario,
                                           horizon = horizon)
   expect_equal_to_reference(tmp, "regressionOutput/radiation/KNMI14_WL_2085_rsds.rds")
 })